Refactor, moved code to print room descriptions from the main game loop into the Room class
