+++ src/room.py
@@ -1,6 +1,17 @@
 # Implement a class to hold room information. This should have name and
 # description attributes.
 
+import textwrap
+from item import LightSource
+
+def print_wrapped(message, width=50):
+    """ Prints a message using the textwrap module """
+    wrapper = textwrap.TextWrapper(width) 
+    message = wrapper.wrap(message) 
+    for line in message: 
+        print(line)
+
+
 class Room():
     """
     The Room class, handles the room name and description, and keeps track of
@@ -42,3 +53,28 @@ class Room():
         
         # No item with that name found
         return None
+    
+    def print_description(self, has_light):
+        """
+        Prints the room description to the console
+
+        has_light (bool): pass in True if the player has a light
+        """
+
+        print_wrapped(self.name)
+
+        # Print description if we can see
+        if self.is_light or has_light:
+            print_wrapped(self.description)
+        else:
+            print("It's pitch black in here!")
+
+        if self.is_light or has_light:
+            # Show all items in the room
+            for item in self.items:
+                print("You see a", item.name)
+        else:
+            # Only show light sources
+            for item in self.items:
+                if type(item) == LightSource:
+                    print("You see a", item.name)
